+++ src/main/java/org/example/expert/domain/todo/dto/request/TodoSearchRequest.kt
@@ -3,8 +3,8 @@ package org.example.expert.domain.todo.dto.request
 import java.time.LocalDateTime
 
 data class TodoSearchRequest(
-    val page: Int = 1,
-    val size: Int = 10,
+    val page: Long = 1L,
+    val size: Long = 10,
     val title: String? = null,
     val since: LocalDateTime? = null,
     val until: LocalDateTime? = null,
